TwilioProvider returns class with initialized .client property, + rename some stuff
