+++ src/cloud/createConnection.js
@@ -1,4 +1,4 @@
-import Parse from '../providers/ParseServiceProvider';
+import Parse from '../providers/ParseProvider';
 import stripPhoneNumber from '../utils/stripPhoneNumber';
 
 /**
