+++ src/services/initiate2FA.js
@@ -1,4 +1,4 @@
-import Twilio from '../providers/TwilioServiceProvider';
+import Twilio from '../providers/TwilioProvider';
 
 
 /**
@@ -7,7 +7,7 @@ import Twilio from '../providers/TwilioServiceProvider';
  * @param {Parse.User} user
  */
 const initiate2FA = (auth_code, user) => {
-  return Twilio.messages.create({
+  return new Twilio().client.messages.create({
     body: `Your code for Benji is: ${ auth_code }`,
     from: '+12012560616',
     to: user.phoneNumber,
